doc(bigquery): add code sample for scripting
This code sample initiates a scripting job and then demonstrates how to
fetch the results and the child jobs.

Removes system test for scripting, as it's redundant with the code
sample.
